Render only the interest-list form on /community
The /community page rendered the whole landing section, so the teaser's
"Meet the community" CTA appeared on the very page it points at. The
module offered no interface through which that could differ: one export,
no props, both jobs inside it.

InterestListForm becomes its own module — the card, the fields and the
success swap — and the landing section composes it beside the teaser.
/community renders it alone, under its own heading, at the width the page
chooses. Two callers now vary across the seam, which is what makes the
seam real rather than hypothetical.

CommunityForm loses "use client" as a side effect: with the state gone
into the form, the teaser is a server component again.

The form's tests move to the module that owns the behavior; the landing
section keeps the teaser-link assertion and gains one that the form is
composed in. The MUST FAIL regression from the previous commit becomes a
plain test() and passes.
