+++ src/app/community/page.test.tsx
@@ -27,17 +27,12 @@ test("the interest-list form rides on the page, not just the landing", () => {
   ).toBeDefined();
 });
 
-// MUST FAIL, per CLAUDE.md's rule that a bugfix starts with a regression test
-// that failed first. The gate table's mustFail flag judges a whole gate, and
-// the test gate runs the whole suite, so it cannot express "one test in this
-// suite must fail". test.fails is that at test granularity: it passes only
-// while the body throws, so this commit proves the defect and leaves the gate
-// green. The fix commit turns it into a plain test().
-test.fails("the page does not offer a link to itself", () => {
+test("the page does not offer a link to itself", () => {
   render(<Community />);
 
-  // The landing teaser rides in on CommunityForm, and its CTA points at
-  // /community — the page the reader is already on.
+  // This page is what the landing teaser's "Meet the community" points at,
+  // so that link must not ride along here. Regression: it did, until the
+  // form became its own module.
   expect(screen.queryByRole("link", { name: /meet the community/i })).toBe(
     null,
   );
